requirements/test.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    make upgrade
#
asgiref==3.7.2
    # via
    #   -r requirements/base.txt
    #   django
astroid==3.0.3
    # via pylint
backports-zoneinfo==0.2.1
    # via
    #   -r requirements/base.txt
    #   django
coverage==7.4.1
    # via -r requirements/test.in
ddt==1.7.1
    # via -r requirements/test.in
dill==0.3.8
    # via pylint
    # via
    #   -c requirements/constraints.txt
    #   -r requirements/base.txt
    #   django-crum
    #   django-mysql
    #   djangorestframework
    #   jsonfield
    #   openedx-filters
django-crum==0.7.9
    # via -r requirements/base.txt
django-fake-model==0.1.4
    # via -r requirements/test.in
django-mysql==4.12.0
    # via -r requirements/base.txt
djangorestframework==3.14.0
    # via -r requirements/base.txt
edx-opaque-keys[django]==2.5.1
    # via
    #   -r requirements/base.txt
    #   edx-opaque-keys
isort==5.13.2
    # via pylint
jsonfield==3.1.0
    # via -r requirements/base.txt
mccabe==0.7.0
    # via pylint
mock==5.1.0
    # via -r requirements/test.in
openedx-filters==1.6.0
    # via -r requirements/base.txt
path==16.10.0
    # via path-py
path-py==12.5.0
    # via -r requirements/test.in
pbr==6.0.0
    # via
    #   -r requirements/base.txt
    #   stevedore
platformdirs==4.2.0
    # via pylint
pycodestyle==2.11.1
    # via -r requirements/test.in
pylint==3.0.3
    # via -r requirements/test.in
pymongo==3.13.0
    # via
    #   -r requirements/base.txt
    #   edx-opaque-keys
pytz==2024.1
    # via
    #   -r requirements/base.txt
    #   djangorestframework
six==1.16.0
    # via -r requirements/base.txt
sqlparse==0.4.4
    # via
    #   -r requirements/base.txt
    #   django
stevedore==5.1.0
    # via
    #   -r requirements/base.txt
    #   edx-opaque-keys
testfixtures==7.2.2
    # via -r requirements/test.in
tomli==2.0.1
    # via pylint
tomlkit==0.12.3
    # via pylint
typing-extensions==4.9.0
    # via
    #   -r requirements/base.txt
    #   asgiref
    #   astroid
    #   edx-opaque-keys
    #   pylint
